+++ app/Providers/AppServiceProvider.php
@@ -17,6 +17,10 @@ class AppServiceProvider extends ServiceProvider
         // Set the default string length for database migrations.
         // @see https://laravel.com/docs/5.4/migrations#indexes
         Schema::defaultStringLength(191);
+
+        // Configure the Bcrypt hasher to use a cost factor
+        // of 12 instead of the default value of 10.
+        app('hash')->setRounds(12);
     }
 
     /**