+++ database/factories/Location.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Partner;
 use App\Location;
 use Faker\Generator as Faker;
 
@@ -7,5 +8,8 @@ use Faker\Generator as Faker;
 $factory->define(Location::class, function (Faker $faker) {
     return [
         'name' => "{$faker->company} {$faker->city}",
+        'partner_id' => function () {
+            return factory(Partner::class)->create()->id;
+        }
     ];
 });