+++ database/migrations/2017_07_15_215105_create_partners_table.php
@@ -19,6 +19,13 @@ class CreatePartnersTable extends Migration
             // An unsigned integer with autoincrement.
             $table->increments('id');
 
+            // Optional foreign key.
+            // This references the team member who made this partner sign the
+            // official documents. This is optional but highly recommended.
+            // Note: this just adds the column. The foreign constraint itself
+            //       is added via the migration for the `team_members` table.
+            $table->unsignedInteger('endorser_team_member_id')->nullable();
+
             // The name of the partner.
             $table->string('name');
 
@@ -27,10 +34,10 @@ class CreatePartnersTable extends Migration
             // Example : 'Du côté de chez Poje' → 'Poje (du côté de chez)'
             $table->string('name_sort')->nullable();
 
-            // The slug is a simplified version of the name that is
-            // both people-friendly and computer-friendly, for use
-            // in places such as URLs or filenames.
-            $table->string('slug');
+            // The slug is a simplified version of the name that is both people-
+            // friendly and computer-friendly, for use in places such as URLs
+            // or filenames. Each partner must have a different one.
+            $table->string('slug')->unique();
 
             // The type of business (ASBL, SPRL, indépendant en
             // personne physique, etc.)