+++ resources/views/public/project.blade.php
@@ -3,11 +3,214 @@
 @section('title', 'Le projet')
 
 @section('content')
-
-    {{-- <h2 class="partner-list-main-heading">Le projet</h2> --}}
-
-    <div class="le-projet">
-        <img src="{{ asset('img/under_construction--blue.svg') }}" alt="">
-        <p>Page en construction… Elle arrivera dès que possible :-)</p>
+<div class="long-text">
+    <div class="long-text__table-of-contents">
+        <h2>Sommaire</h2>
+        <ol>
+            <li>
+                <a href="#les-origines-du-val-heureux">
+                    Les origines du Val’heureux
+                </a>
+            </li>
+            <li>
+                <a href="#structure-du-projet">
+                    Structure du projet
+                </a>
+            </li>
+            <li>
+                <a href="#elargissement-du-reseau-de-partenaires">
+                    Élargissement du réseau de partenaires
+                </a>
+            </li>
+            <li>
+                <a href="#communication-et-promotion-du-val-heureux">
+                    Communication et promotion du Val’heureux
+                </a>
+            </li>
+            <li>
+                <a href="#partenariat-avec-la-ville-de-liege">
+                    Partenariat avec la ville de Liège
+                </a>
+            </li>
+            <li>
+                <a href="#questions-frequemment-posees-par-les-commercants">
+                    Questions fréquemment posées par les commerçants
+                </a>
+            </li>
+        </ol>
     </div>
+
+    <h2 id="les-origines-du-val-heureux">Les origines du Val’heureux</h2>
+
+    <h3>Origine du projet</h3>
+
+    <p>L’idée de créer un bon de soutien à l’économie locale de la région liégeoise est née en 2011 à l’initiative de quelques citoyens actifs au sein de Liège en Transition ou liés au réseau Financité. Leur but : favoriser l’économie réelle, locale et éthique.</p>
+
+    <p>Ce groupe d’une dizaine de personnes a élaboré son projet dans le cadre des actions menées par l’ASBL Barricade, par Liège en Transition et par les promoteurs de la Ceinture aliment-terre liégeoise (<a href="https://www.catl.be">www.catl.be</a>). Début 2014, une ASBL a été fondée et, en juin de la même année, les premiers billets de Valeureux sont apparus dans les portefeuilles de la Cité ardente.</p>
+
+    <p>Début 2017, différents groupes de citoyen·ne·s de la province de Liège en réflexion sur l’économie locale et les circuits courts ont commencé à s’intéresser à la monnaie citoyenne. Ces groupes sont issus d’initiatives de transition (comme Comblain-au-Pont en Transition), de dynamiques de jeunes citoyens (Génér’action de Huy-Waremme) et d’associations comme De Bouche à Oreille (<a href="http://www.dbao.be">www.dbao.be</a>). Après quelques soirées de formation et de réflexion, ces différents groupes ont décidé d’unir leurs efforts pour travailler sur une nouvelle version de la monnaie citoyenne liégeoise. Ainsi, le Valeureux évoluait en Val’heureux, pour mieux soutenir les circuits courts d’une bonne partie de la province de Liège.</p>
+
+    <p>Ce sont donc désormais cinq comités de citoyen·ne·s, rassemblés au sein de l’ASBL le Val’heureux, qui assurent la promotion et la gestion du Val’heureux. Ces groupes sont répartis dans les régions de Liège, Huy-Hesbaye-Condroz, Ourthe-Amblève, Verviers et Pays de Herve.</p>
+
+    <h3>Contexte du projet</h3>
+
+    <p>Le Val’heureux s’inscrit dans le vaste mouvement planétaire des « monnaies complémentaires ». On estime qu’il en existe aujourd’hui plus de 5000 dans le monde. En Belgique francophone, il y a, actuellement, six autres « monnaies complémentaires » en circulation : l’Épi lorrain à Virton, les Blés à Grez-Doiceau, le Talent à Ottignies/Louvain-La-Neuve, le Voltí à Rocherfort-Ciney-Marche, le Lumsou à Namur et le Sous-rire à Malmédy et le Ropi à Mons. D’autres projets sont également en voie de construction ailleurs en Wallonie et à Bruxelles. Enfin, le Toreke, soutenu par la ville de Gand, est, pour les promoteurs du Val’heureux, un modèle très inspirant.</p>
+
+    <p>Certain·e·s pourraient être surpris·es qu’une poignée de citoyens prétendent créer une « monnaie ». En effet, l’émission de la « monnaie » est une prérogative aujourd’hui réservée à la Banque Centrale européenne et à la Banque nationale de Belgique. Ce que nous appelons ici « monnaie » est donc, d’un point de vue légal, un « bon de soutien à l’économie locale ». Cette mention figure sur tous les documents officiels.</p>
+
+    <p>Si les promoteurs et promotrices du Val’heureux s’investissent dans la création et la gestion d’un bon de soutien à l’économie locale, c’est parce que, comme Bernard Lietaer (économiste et ancien haut fonctionnaire de la Banque nationale de Belgique), l’écrit, <em>« ils sont convaincus que l’euro ne suffit pas ! Il manque un outil économique destiné à stimuler exclusivement les échanges locaux, entièrement dédié au service de l’économie réelle, capable de renforcer l’emploi non délocalisable »</em>. En résumé, sa finalité consiste à orienter les comportements de consommation, soutenir les petits entrepreneurs et rapprocher les acteurs de l’économie locale pour répondre aux enjeux d’un développement plus durable et solidaire.</p>
+
+    <p>La région liégeoise compte de nombreux entrepreneurs qui nourrissent le tissu économique local. Ils et elles travaillent au quotidien dans l’esprit d’un développement économique respectueux des personnes et de l’environnement. L’objectif du Val’heureux est donc de renforcer ces projets et leur clientèle.</p>
+
+    <p>Pour ses promoteurs et promotrices, le Val’heureux constitue donc une alternative à la spirale infernale d’une certaine mondialisation économique qui déstructure sans cesse les économies locales, accapare les richesses, distend les liens sociaux et détruit l’environnement, source de toute prospérité.</p>
+
+    <p>Enfin, via sa structure en ASBL, le Val’heureux est un outil transparent et démocratique, fruit d’une gestion citoyenne et participative.</p>
+
+    <h2 id="structure-du-projet">Structure du projet</h2>
+
+    <h3>Objectifs du Val’heureux</h3>
+
+    <p>Le Val’heureux est un complément régional à l’euro. Le simple fait de transformer un euro en Val’heureux crée de la richesse qui circule dans un territoire donné, celui des bassins de vie en région liégeoise. Dès lors, le Val’heureux poursuit plusieurs objectifs :</p>
+
+    <ul>
+        <li>renforcer et valoriser l’économie locale et les circuits courts ;</li>
+        <li>favoriser l’utilisation de biens et de services socialement responsables ;</li>
+        <li>soutenir les initiatives respectueuses de l’environnement ;</li>
+        <li>promouvoir en particulier la souveraineté alimentaire et économique ;</li>
+        <li>créer du lien social sur base locale tout en facilitant les échanges.</li>
+    </ul>
+
+    <p>Ainsi, le Val’heureux permet de retenir et de mieux faire circuler la richesse créée par les petits entrepreneurs, les citoyens et les acteurs de l’économie locale et éthique.</p>
+
+    <h3>Le Val’heureux en pratique</h3>
+
+    <p>Le Val’heureux est un bon de soutien à l’économie locale pour les bassins de vie de la région liégeoise. Il est partagé par un réseau limité d’utilisateurs. La parité entre l’euro et le Val’heureux est fixe : 1 Val’heureux = 1 euro. Le Val’heureux se présente sous la forme de bons sécurisés imprimés sur papier. Il y a six coupures de billets, respectivement d’une valeur de 0 (zéro), 0,5 (un demi), 1 (un), 5 (cinq), 10 (dix) et 20 (vingt) Val’heureux.</p>
+
+    <p>Ces billets ont été créés par la graphiste Colienne Roberti. Les promoteurs du Val’heureux n’ont rien prévu pour l’utilisation de la coupure de zéro Val’heureux, pariant sur l’imagination et l’ingéniosité de leurs concitoyen·ne·s pour en faire bon usage.</p>
+
+    <p>Chaque coupure est liée à un thème :</p>
+
+    <ul>
+        <li>le lien social pour le billet de zéro ;</li>
+        <li>l’environnement pour le billet d’un demi ;</li>
+        <li>la culture pour le billet de un ;</li>
+        <li>la souveraineté alimentaire pour le billet de cinq ;</li>
+        <li>la liberté et la démocratie pour le billet de dix ;</li>
+        <li>la solidarité pour le billet de vingt.</li>
+    </ul>
+
+    <p>Chaque coupure possède un recto unique et un verso qui, lui, est spécifique à chaque région (Liège, Huy-Hesbaye-Condroz, Verviers, Ourthe-Amblève et Pays de Herve). Ces versos reprennent un paysage, un monument ou un symbole local. Il y a donc trente billets différents (6 coupures, chacune avec 5 faces régionales différentes), dont la cohérence est assurée par une même charte graphique et qui peuvent circuler indifféremment partout où le Val’heureux est accepté.</p>
+
+    <h3>Les acteurs du système</h3>
+
+    <p>Toute personne physique ou morale peut utiliser des Val’heureux. Deux rôles différents peuvent être joués : celui d’« utilisateur·trice », ou celui de « prestataire partenaire ».</p>
+
+    <h4>L’utilisateur·trice</h4>
+
+    <p>Est utilisatrice toute personne qui participe à une transaction commerciale en Val’heureux. Elle peut devenir « membre adhérent » de l’ASBL pour marquer son soutien au système. En posant ce choix, elle s’identifie et est répertoriée dans un registre.</p>
+
+    <p>Les « membres adhérents » peuvent devenir « membres effectifs » de l’ASBL si leur demande est acceptée et s’ils s’acquittent d’une cotisation dont le montant est fixé par l’assemblée générale sur proposition du conseil d’administration.</p>
+
+    <h4>Le prestataire partenaire</h4>
+
+    <p>Les prestataires partenaires sont des indépendants, des entreprises, des associations ou toute autre organisation qui vendent un bien ou un service et qui acceptent le Val’heureux comme moyen de paiement. Ils ont signé la charte du Val’heureux et adhèrent donc à ses principes. Tous les prestataires partenaires sont membres effectifs de l’assemblée générale de l’ASBL « le Val’heureux », où leurs idées, leurs avis et leurs voix sont, bien entendu, pris en compte. Les prestataires partenaires s’acquittent d’une cotisation annuelle fixée au minimum à 25€.</p>
+
+    <h3>Dépenser les Val’heureux</h3>
+
+    <p>Les Val’heureux peuvent être dépensés chez tous les prestataires partenaires. Ceux-ci figurent dans l’annuaire du Val’heureux publié notamment sur son site web. Durant les événements partenaires, le Val’heureux peut être utilisé en lieu et place des tickets.</p>
+
+    <p>Lors d’un paiement en Val’heureux, la monnaie est rendue préférentiellement en Val’heureux. Pour la petite monnaie (cents) ou lorsque les coupures nécessaires ne sont pas disponibles, le change est rendu en euros.</p>
+
+    <p>Lorsqu’il compte sa caisse en fin de journée, le prestataire partenaire comptabilise les euros et les Val’heureux en les intégrant indifféremment à son chiffre d’affaires.</p>
+
+    <h3>Principe de gestion de l’ASBL « Le Val’heureux »</h3>
+
+    <p>Les euros convertis en Val’heureux sont versés sur une réserve de contrepartie (un compte ouvert auprès de la banque Triodos), qui ne sert que dans ce but. La gestion quotidienne et les frais de fonctionnement de l’ASBL sont financés par un compte bancaire séparé.</p>
+
+    <h3>Secteurs d’activité concernés par le Val’heureux</h3>
+
+    <p>Le Val’heureux s’applique à tous les secteurs d’activité. Toutefois, la priorité est donnée au secteur alimentaire.</p>
+
+    <p>La zone géographique d’action du Val’heureux inclut Liège, Verviers, Huy, la Hesbaye, le Condroz, Ourthe-Amblève et le Pays de Herve. Le Val’heureux tient compte des boucles économiques locales et de l’évolution de la Ceinture aliment-terre liégeoise (<a href="https://www.catl.be">www.catl.be</a>), en voie de construction.</p>
+
+    <h3>L’ASBL « Le Val’heureux »</h3>
+
+    <p>Durant le premier trimestre de l’année 2014, les statuts de l’ASBL ont été rédigés. Ils ont été validés par le tribunal de commerce de Liège le 19 mars 2014. Les statuts ont été adaptés en octobre 2017. Ils sont disponibles notamment via le site web du Val’heureux (<a href="https://valheureux.be">valheureux.be</a>).</p>
+
+    <h3>Respect des objectifs et promotion du dispositif</h3>
+
+    <p>L’ASBL « le Val’heureux » gère et développe le système mis en place. Elle veille au respect de la finalité sociale et aux objectifs du projet. Elle promeut l’utilisation des Val’heureux et encourage l’adhésion de nouveaux prestataires partenaires. Elle assure la promotion du système auprès du grand public et des commerçants. Elle encourage l’utilisation des Val’heureux tout au long des filières.</p>
+
+    <p>Outre les prestataires partenaires, sont membres effectifs de l’ASBL, les particuliers dont la demande d’adhésion, soumise à l’assemblée générale par le conseil d’administration, a été acceptée.</p>
+
+    <h3>Partenariat avec le réseau « Financité »</h3>
+
+    <p>L’ASBL « le Val’heureux » a signé une convention de partenariat avec le réseau « Financité ». Son statut de « groupe local » lui permet de rester en contact et de développer des collaborations avec les autres groupes actifs dans le développement d’une monnaie citoyenne en Wallonie, à Bruxelles et également au niveau international. L’ASBL « le Val’heureux » profite également d’un soutien logistique et humain mis à disposition par Financité. Grâce aux heures d’éducation permanente réalisées par les volontaires du Val’heureux et valorisées par Financité, l’ASBL bénéficie également de quelques moyens financiers via le réseau.</p>
+
+    <h2 id="elargissement-du-reseau-de-partenaires">Élargissement du réseau de partenaires</h2>
+
+    <p>En novembre 2017, le Val’heureux compte plus de 150 partenaires. La liste complète des partenaires est actualisée de manière régulière et est accessible sur le site web du Val’heureux et sur sa page Facebook.</p>
+
+    <p>La liste actualisée des comptoirs de change peut également être consultée sur le site.</p>
+
+    <h2 id="communication-et-promotion-du-val-heureux">Communication et promotion du Val’heureux</h2>
+
+    <h3>Sur internet</h3>
+
+    <p>Le site internet <a href="https://valheureux.be">valheureux.be</a> est régulièrement mis à jour, notamment à chaque adhésion d’un nouveau partenaire.</p>
+
+    <p>On y trouve l’actualité du Val’heureux, l’annuaire des partenaires, toute la documentation relative au fonctionnement du Val’heureux et de l’ASBL, ainsi que la charte. Il est également possible d’envoyer un message à {!! Html::mailto('info@valheureux.be') !!}.be ou via le formulaire en ligne. Les courriels envoyés à {!! Html::mailto('info@valheureux.be') !!}.be sont traités dans les 24 heures.</p>
+
+    <p>La page Facebook du Val’heureux est également alimentée régulièrement. Elle permet de faire connaître le Val’heureux à un plus grand nombre de personnes et surtout de faire connaître l’actualité du Val’heureux et des commerçants partenaires. La page Facebook compte actuellement plus de 2000 <em>followers</em>. Une page Facebook Val’heureux pour chaque comité régional est en cours de création.</p>
+
+    <h3>Matériel de promotion et signalétique</h3>
+
+    <p>Différents documents et supports de promotion ont été élaborés, notamment un flyer de présentation du Val’heureux et un dépliant répertoriant tous les commerçants partenaires.</p>
+
+    <p>Ces documents sont actualisés régulièrement et sont disponibles sur les comptoirs des commerçants partenaires et sont distribués lors d’événements.</p>
+
+    <p>Les prestataires partenaires disposent également d’autocollants « Ici circule le Val’heureux », qu’ils peuvent apposer sur leur vitrine ou à tout autre endroit approprié.</p>
+
+    <h2 id="partenariat-avec-la-ville-de-liege">Partenariat avec la ville de Liège</h2>
+
+    <p>Au printemps dernier, M. Demeyer, bourgmestre de Liège, a soumis au vote du conseil communal une motion visant à soutenir le développement du Val’heureux. Grâce à cet appui officiel, de nouveaux partenariats ont vu le jour : collaboration avec le guichet de l’énergie, l’office du tourisme ou encore le commerce liégeois et diverses autres structures d’organisation d’événements en Cité ardente.</p>
+
+    <h2 id="questions-frequemment-posees-par-les-commercants">Questions fréquemment posées par les commerçants</h2>
+
+    <h3>Quels sont les risques financiers ?</h3>
+
+    <p>Aucun. Chaque euro changé en Val’heureux est placé en sécurité sur un compte Triodos appelé réserve de contrepartie. Si le système devait se casser la figure, chaque commerçant ou citoyen qui détient des Val’heureux pourrait les rééchanger contre des euros.</p>
+
+    <h3>Quel impact sur ma comptabilité ?</h3>
+
+    <p>Le Val’heureux ne change rien à la comptabilité du commerce. D’un point de vue comptable, quand un V entre dans la caisse, c’est comme si un € entrait dans la caisse. La parité de 1 € = 1 V a été choisie pour faciliter la gestion du V et surtout pour faciliter la vie des commerçants et des producteurs partenaires. Il n’y a donc aucun souci avec la TVA ou les impôts. Il est à noter que, contrairement à certaines idées reçues, la parité 1 € = 1 V n’enlève rien à l’efficacité et aux vertus du Val’heureux.</p>
+
+    <h3>Et si j’ai trop de Val’heureux en caisse ?</h3>
+
+    <p>Un tel projet met du temps à se construire. Pour l’instant, un commerçant qui a trop de Val’heureux en caisse et qui ne sait pas comment les faire circuler peut demander à l’ASBL le Val’heureux de lui reconvertir gratuitement ses V contre des €. Plus tard, quand davantage de boucles économiques seront établies et que les commerçants auront des possibilités de payer leurs fournisseurs en V, l’assemblée générale, dont les commerçants partenaires font partie, pourra éventuellement décider de faire payer des frais de change (uniquement pour des reconversions de Val’heureux en euros) pour décourager les commerçants de faire sortir des V du système. C’est l’assemblée générale qui pourra aussi décider également d’éventuelles exceptions pour certains secteurs (exemple, les bijouteries, parce que les bijoutiers ont peu de fournisseurs locaux – il faudra néanmoins examiner les autres possibilités de faire circuler le V).</p>
+
+    <p>Aujourd’hui, le Val’heureux est toujours dans sa phase de démarrage (lancer une monnaie citoyenne et reconstruire un système économique local prend du temps !) et il n’y a encore que peu de producteurs dans le système, donc aucun frais de reconversion n’est envisagé. De plus, une version « électronique » du Val’heureux est en chantier (PC banking, smartphone et SMS). Il permettra d’encore mieux faire circuler la monnaie. Des verrous seront levés. La question des frais sera alors peut-être posée à l’assemblée générale.</p>
+
+    <h3>J’ai peur d’être submergé de Val’heureux !</h3>
+
+    <p>Le V est lancé depuis juin 2014 et les commerçants ne se plaignent pas. De temps en temps, des commerçants regrettent de ne pas voir davantage de gens payer en Val’heureux. C’est normal. Le V est toujours dans sa phase de démarrage. En comparaison avec la taille de l’économie régionale, il y a encore peu de V en circulation. Aujourd’hui, l’important est de créer le réseau et de soutenir la mise en place d’une économie plus durable qui soutient les commerces de proximité et les petits producteurs régionaux. Quand viendra le moment où plus de V seront dépensés dans les commerces, il y aura aussi plus de possibilités de les dépenser chez des fournisseurs, et plus de clients voudront aussi récupérer des V. À ce moment, l’afflux de V ne sera donc pas un problème. Au contraire, le V prendra tout son sens.</p>
+
+    <h3>Ai-je mon mot à dire ?</h3>
+
+    <p>Bien entendu. Le V est un outil économique au service de l’économie locale et du développement durable mais est aussi et surtout un outil au service de la souveraineté économique. Autrement dit, le V est une manière pour les citoyens et les commerçants de se réapproprier la gestion de l’économie locale. Les promoteurs du V souhaitent qu’un maximum de commerçants et de producteurs partenaires et d’utilisateurs s’investissent dans la réflexion et le développement du V. Les commerçants partenaires font partie de l’assemblée générale, pourront, s’ils le désirent, être élus au conseil d’administration, et peuvent apporter toutes les réflexions et critiques et proposer toutes les améliorations qu’ils souhaiteraient. Le V fonctionnera encore mieux si les commerçants et les citoyens se l’approprient.</p>
+
+    <h3>Quels sont les avantages pour le commerçant ?</h3>
+
+    <ul>
+        <li>Visibilité et promotion faite par l’ASBL « le Val’heureux », notamment via le répertoire des prestataires partenaires, les flyers, le site, la page Facebook, les passages dans la presse et via les événements auxquels l’ASBL participe ou organisés par elle ;</li>
+        <li>affirmation de l’identité (développement durable et local) et différenciation par rapport à la grande distribution ;</li>
+        <li>fidélisation de la clientèle ;</li>
+        <li><p>possibilité de simplification des systèmes de « cartes de fidélité » ;</p>
+        Notons qu’il serait intéressant que les commerçants, au lieu de faire une ristourne pour remercier la fidélité de leur client, rendent l’équivalent de cette ristourne en Val’heureux. C’est bon pour le client, qui bénéficie toujours d’un cadeau, c’est bon pour le système Val’heureux puisque ça stimule la circulation de la monnaie et, enfin, c’est bon pour le commerçant qui ne perd pas, purement et simplement, le montant de sa ristourne puisque les Val’heureux offerts par le commerçant, ou par un autre membre du système, vont revenir au commerçant.</li>
+        <li>intégration à un réseau de clients partagés avec les autres partenaires du système et un élargissement de la clientèle ;</li>
+        <li>mise en réseau avec d’autres partenaires (dont des producteurs ou des fournisseurs) qui privilégient les objectifs du Val’heureux ;</li>
+        <li>le Val’heureux a pour objectif de dynamiser le commerce local et donc de soutenir les petits entrepreneurs ;</li>
+        <li>profiter des actions lancées par l’ASBL le Val’heureux comme par exemple l’enveloppe cadeau lors des fêtes de fin d’année (V dépensés dans le réseau, évidemment !).</li>
+    </ul>
+</div>
 @endsection